use FeedForward class instead of another MLP
